LPS-125551 Delegate removeAppender and get configuration file content(loggerName，level) to Log4JConfigurator.
